# template/services/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/users/grant": {
      "post": {
        "operationId": "grant",
        "summary": "Grant",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccess"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserGrant"
              }
            }
          }
        }
      }
    },
    "/users/access": {
      "get": {
        "operationId": "resolve",
        "summary": "Resolve",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccess"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "external_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "title": "User",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "default": "inactive"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "UserChannel": {
        "type": "object",
        "title": "UserChannel",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "user_id": {
            "type": "integer",
            "readOnly": true
          },
          "channel": {
            "type": "string",
            "minLength": 1
          },
          "external_id": {
            "type": "string",
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "user_id",
          "channel",
          "external_id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "UserGrant": {
        "type": "object",
        "title": "UserGrant",
        "properties": {
          "channel": {
            "type": "string",
            "minLength": 1
          },
          "external_id": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "channel",
          "external_id"
        ],
        "additionalProperties": false
      },
      "UserAccess": {
        "type": "object",
        "title": "UserAccess",
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "channel": {
            "type": "string"
          },
          "external_id": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "status",
          "channel",
          "external_id"
        ],
        "additionalProperties": false
      },
      "CommandReceived": {
        "type": "object",
        "title": "CommandReceived",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "command",
          "args",
          "user_id",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "CommandReceivedCreate": {
        "type": "object",
        "title": "CommandReceivedCreate",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "command",
          "args",
          "user_id"
        ],
        "additionalProperties": false
      }
    }
  }
}