# template/services/backend/src/generated/routers/users.py
##############################################################################
# THIS FILE IS GENERATED BY THE FRAMEWORK. DO NOT EDIT MANUALLY.
# CHANGES WILL BE OVERWRITTEN.
##############################################################################

# AUTO-GENERATED BY SERVICE-TEMPLATE - DO NOT EDIT
"""FastAPI router for users."""

from __future__ import annotations

from collections.abc import Callable

from fastapi import (
    APIRouter,
    Body,
    Depends,
    Query,
)
from faststream.redis import RedisBroker
from sqlalchemy.ext.asyncio import AsyncSession

from shared.generated.schemas import (
    UserAccess,
    UserGrant,
)

from ..protocols import UsersControllerProtocol


def create_router(
    *,
    get_session: Callable[..., object],
    get_controller: Callable[[], UsersControllerProtocol],
    get_broker: Callable[[], RedisBroker],
) -> APIRouter:
    router = APIRouter(
        prefix="/users",
        tags=["users"],
    )

    @router.post(
        "/grant",
        response_model=UserAccess,
        status_code=200,
    )
    async def grant(
        payload: UserGrant = Body(...),  # noqa: B008
        session: AsyncSession = Depends(get_session),  # noqa: B008
        controller: UsersControllerProtocol = Depends(get_controller),  # noqa: B008
        broker: RedisBroker = Depends(get_broker),  # noqa: B008
    ) -> UserAccess:
        result = await controller.grant(
            session=session,
            payload=payload,
        )
        await broker.publish(result, "user_granted")
        return result

    @router.get(
        "/access",
        response_model=UserAccess,
        status_code=200,
    )
    async def resolve(
        channel: str = Query(...),  # noqa: B008
        external_id: str = Query(...),  # noqa: B008
        session: AsyncSession = Depends(get_session),  # noqa: B008
        controller: UsersControllerProtocol = Depends(get_controller),  # noqa: B008
    ) -> UserAccess:
        result = await controller.resolve(
            session=session,
            channel=channel,
            external_id=external_id,
        )
        return result

    return router