# template/shared/shared/generated/schemas.py
##############################################################################
# THIS FILE IS GENERATED BY THE FRAMEWORK. DO NOT EDIT MANUALLY.
# CHANGES WILL BE OVERWRITTEN.
##############################################################################

# generated by datamodel-codegen:
#   filename:  <stdin>

from __future__ import annotations

from enum import StrEnum
from typing import Any

from pydantic import AwareDatetime, BaseModel, ConfigDict, RootModel, constr


class Status(StrEnum):
    active = "active"
    inactive = "inactive"


class User(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    id: int
    status: Status = "inactive"
    created_at: AwareDatetime
    updated_at: AwareDatetime


class UserChannel(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    id: int
    user_id: int
    channel: constr(min_length=1)
    external_id: constr(min_length=1)
    created_at: AwareDatetime
    updated_at: AwareDatetime


class UserGrant(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    channel: constr(min_length=1)
    external_id: constr(min_length=1)


class UserAccess(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    user_id: int
    status: Status
    channel: str
    external_id: str


class CommandReceived(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    command: str
    args: list[str]
    user_id: int
    timestamp: AwareDatetime


class CommandReceivedCreate(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    command: str
    args: list[str]
    user_id: int
    timestamp: AwareDatetime | None = None